+++ main.cjs
@@ -179,7 +179,7 @@ function createMenuTemplate() {
         { role: "zoomout" },
         { role: "togglefullscreen" },
         { type: "separator" },
-        { role: "toggledevtools" },
+        //{ role: "toggledevtools" },
       ],
     },
     {
@@ -191,7 +191,7 @@ function createMenuTemplate() {
             dialog.showMessageBox({
               type: 'info',
               title: 'EasyEdit',
-              message: 'EasyEdit v1.0.9 \n\n EasyEdit is an easy markdown editor that allows you to write MarkDown (MD) and preview it in real-time. You can save, load .md files and export to HTML & PDF. \n\nDeveloped by: Ricardo Wagemaker <wagemra@gmail.com> \nGitHub: https://github.com/gcclinux/EasyEdit \nLicense: MIT\n',
+              message: 'EasyEdit v1.1.0 \n\n EasyEdit is an easy markdown editor that allows you to write MarkDown (MD) and preview it in real-time. You can save, load .md files and export to HTML & PDF. \n\nDeveloped by: Ricardo Wagemaker <wagemra@gmail.com> \nGitHub: https://github.com/gcclinux/EasyEdit \nLicense: MIT\n',
               buttons: ['OK']
             });
           },