+++ CHANGELOG.md
@@ -3,13 +3,13 @@
 EasyEdit is an easy markdown editor that allows you to write Markdown (MD) and preview it in real-time. You can save, load .md files and export to HTML & PDF. All notable changes to this project will be documented in this file.
 
 ## Latest version
-- 1.0.9 - To improve development of saveAsPDF function the code has been moved into separate file saveAsPDF.tsx
+- 1.1.0 - Additional graph plaintext and flowchart
 
 ## Latest built version
 - 1.0.8 - Windows (exe), Linux (AppImage)
 - - EasyEdit-x86_64.AppImage
 - - EasyEdit-arm64.AppImage
-- - EasyEdit.Setup.1.0.6.exe
+- - EasyEdit.Setup.1.0.8.exe
 
 ## Bug Fix
 - 1.0.1 - Added detect-port in case App starts and default port is already in use by another App
@@ -24,3 +24,4 @@ EasyEdit is an easy markdown editor that allows you to write Markdown (MD) and p
 - 1.0.6 - Focused on a Linux installation script
 - 1.0.8 - Implemented HTML code format and Export Markdown to HTML support
 - 1.0.9 - To improve development of saveAsPDF function the code has been moved into separate file saveAsPDF.tsx
+- 1.1.0 - Additional graph plaintext and flowchart